requirements.txt
pytest==7.1.2
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
xlrd==2.0.1
PyMySQL~=1.0.2
DBUtils~=3.0.2
allure-pytest==2.10.0